+++ showcase/robot-arm/tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/animationcontrols.tsx","./src/components/kukaarm.tsx","./src/components/robotpositioncontroller.tsx","./src/stores/useanimationstore.ts","./src/stores/usemodularstore.ts","./src/stores/userobotarmstore.ts","./src/utils/gcodeparser.ts"],"version":"5.8.3"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/animationcontrols.tsx","./src/components/kukaarm.tsx","./src/components/robotpositioncontroller.tsx","./src/stores/usemodularstore.ts","./src/stores/userobotanimationstore.ts","./src/utils/gcodeparser.ts"],"version":"5.8.3"}
